+++ CHANGELOG.md
@@ -7,14 +7,18 @@
   - Added `injectLatencyInfo()` function that takes an array of triples and injects them into the chart just as if they had been received "in real time"
   - Added `getLatencyMonitorStats()` that retrieves `{ MOSQuality, PacketLoss, Latency, Jitter }` with arrays of four values that mirror those in the Latency Monitor.
 
+---
+
+## Version 0.2.10 - 2025-12-013
+
+- Improved logging surrounding connection start/end; removed redundant logger.info() & console.log()
+- Add "reason" code to `/api/stats`
 - Set fixed-width fields for ms values in Latency Monitor panel
 - Add new API `/api/beacon` that is triggered when the client changes, shows/hides, etc.
   the Cutie tab; triggered by _lifecycle-beacon.ts_
 - More documentation/comments for `api/webrtc`
 - Slightly tuned up _deploy.sh_ script
 - Bump version to 0.2.10
-
----
 
 ## Version 0.2.9 - 2025-12-03